# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Pet Helper API Documentation
  version: 1.0.0
  description: Tài liệu API cho dự án Pet Helper. Hỗ trợ xác thực bằng JWT.
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/orders:
    post:
      summary: Create an order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer: { type: object }
                items:
                  type: array
                  items: { type: object }
      responses:
        200: { description: Order created }
    get:
      summary: Get all orders (Admin)
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: List of all orders }
...
  /api/v1/products:
    get:
      summary: Get all products
      tags: [Products]
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        200: { description: List of products }
    post:
      summary: Add product
      tags: [Products]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: Product added }
...
  /api/v1/reviews:
    post:
      summary: Add a review
      tags: [Reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id: { type: integer }
                user_id: { type: integer }
                rating: { type: integer }
                comment: { type: string }
      responses:
        200: { description: Review added }
...